backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.10
requests==2.32.3
python-decouple==3.8
pydantic==2.9.2
